Color fixes and breeding changes
in Daycare.asm, made it so when you deposit a pokemon it immediately checks for an egg step.
In happinessegg.asm, made it so that an egg check is performed every 10 steps.
CGB_LAYOUTS: Fixed the palette issue if your DV byte was FF.

Evo anim: Fixed broken graphics when playing sprite anim

breeding: set color palettes correctly when hatching eggs